avance actividad modem

+++ K32L2B31A_Project/iot_sdk/iot_sdk_peripherals/sensor_luz.c
@@ -53,7 +53,7 @@
  * Public Source Code
  ******************************************************************************/
 
-float SenLuzObtenerDatoLux(void){
+uint32_t SenLuzObtenerDatoLux(void){
 	SenLuzIniciarCap();
 	SenLuzEsperarResult();
 	uint32_t resultadoADC;
